Create the metadata presenter Rails engine
We are going to use this as a Rails engine instead of a standard gem as
this presenter is going to be doing a lot of logic and routing. We felt
it makes more sense for it to be a mini 'application' inside whatever
containing app that will use it.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Matt Tei <matt.tei@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>