Allow rendering of templates based on page type
This follows a similar structure that fb-components takes in order to
identify which template to use. In the fb-components the `_type`
represents the location of the schema and template relative to the
specifications folder. So all things related to the start page with
 `_type: page.start` is found in specifications/page/start.

Taking a similar approach for the moment the views folder now has a
`page` directory inside of which is the `start.html.erb` template as well
as the `form.html.erb` template.

The form template is a little confusing it has both a page
schema as well as a referenced component schema. The form component
schema is in fact referenced in both the start page schema and the form
page schema.

So this approach expects to look for path in the requested URL and then
extrapolate which page template to use.