Save the data into the session
There are some caveats of using the request.referer:

  The editor uses '/services/:service_id/preview' as base url.

  The problem redirecting to '/text-first' is that this url doesn't
  exist on the editor (only '/services/:service_id/preview/text-first')

  Adding he request.script_name as suggested here:
  https://github.com/rails/rails/issues/7977 makes the redirection
  to work as expected

Also using request.env['PATH_INFO'] (which is 'text-first'
works instead of request.path (which was
'/services/:service_id/preview/text-first')

I tried to add a current page as hidden field but also has some
drawbacks so I rollback.

Remove spring: This is not need on the dummy app because we don't need t
o dummy app to be quick reloading as we don't often change the app, only
the gem

Now the post request expect a page.url which is the current page.

Example:

You are on the first page and submit the form:

        The app will make a POST /reserved/first-page/answers
        and then redirect to /second-page.

Then when you are on the second page and submit the form:

        The app will make a POST /reserved/second-page/answers
        and then redirect to next page.