Add feature spec for the navigation
The test loads the start page add value to fields and go back
to check the value.

Add capybara and site prism gem to use on this kind of specs

We use the current prod complain form metadata with a slight change
on the name page because we don't support multiple question just yet.