Set govuk publishing components environment variables
GOVUK_APP_DOMAIN and  GOVUK_WEBSITE_ROOT are required by Plek which is a
dependency of the govuk_publishing_components gem. This is used by
govspeak, which is actually what we require.

Plek is not something we use but the fact they are not set is stopping
apps that use the presenter from starting.

Temporarily set them until we find an alternative.