Create a common interface for validation
We create the Base validator which should responds to #valid? In order
to be more syntax sugar and near to rails behavior. Also the validate
answer should rely on all validators.

The interface to create validator is as follow:

Each subclass needs to implement the following method:

  invalid_answer?

The Base validator will have #page, #answers & #component attribute

The Base validator will validate and call invalid_answer? as its criteria
to add errors to the page (see RequiredValidator and MaxLengthValidator)

The good thing about it is that is flexible to other validations
to overwrite any method that BaseValidator

Also the default behaviour is to show the custom error message if there
is any on metadata or show the default one from the default metadata.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>