Add change your answer behaviour
We create a reserved route as a post that sets the session to
return to the check your answer and then redirects the user
to the page to change the answer. Once the user changed the
answer and click 'Continue' we verify the session and
redirect the user to the check your answer page.

We create the NextPage class responsible to find which next page
the user should be redirect to. Later it can be add the conditional
pages. Probably we will need the previous page class as well.

Also created a test helper called metadata_fixture(:name) will
look for name.json on spec fixtures