Make sure find page by url can be with or without slashes
This is a temporary solution while we do the definition on how we should
store the urls in the metadata.

Right now when we publish forms to test environment the forms can't
be rendered because the metadata contains the url without the slash
(e.g name) and when the request path info line on pages controller
is called the url to find is '/name' returning nil and not
rendering the page we expected to be rendered.

Do with or without slashes for now solves the issue.

Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>