Single question should have legend or label
On single question we should not show the heading as h1 only.

Quoting from design system:

https://design-system.service.gov.uk/components/text-input/#how-it-works

If you're asking just one question per page as recommended, you can set
the contents of the <label> as the page heading. This is good practice
as it means that users of screen readers will only hear the contents once.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>