Encapsulate answers in separate object
The govuk_design_system_formbuilder gem currently has no way for us to
override the value set on a field for certain components such as number
and radios and checkboxes.

We have previously been relying this functionality with the text and
textarea components in order to present the user with their answer
should they navigate back to a page.

Now we are creating a PageAnswers intermediary object that is
responsible for extracting the users answers and making them available
via a method on the form object that relates to the field id/name (these
are currently the same values).

This does mean that validation responsibility now moves away from the
Page object and into the PageAnswers object.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>