Handle radio button components in metadata
This adds the ability for the presenter to handle the radio buttons
component as well as validate it as a required field.

We added the humanised_title method in order to be able to correctly
choose between using a label or legend property for page
headings/questions, the check your answers page, as well as the key when
creating the errors hash on the page answers object.