Make back link and service name to work in editor preview
When a user clicks on the preview button on the page flow page or the
preview link from the forms page a new tab is opened in their browser.

The preview will function as if it is on the runner and therefore the
back links needed to be adjusted to accommodate this.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>