Allow component title helper method to accept tags and classes
For the multiple question page neither the component legend of label
should be the h1 element nor be styled as extra large. Therefore we need
to be able to allow for a tag and additional classes to be supplied to
the helper in order to flexibly render components on multiple questions
pages.