Add multiple question schemas and default metadata
This adds the multiple questions schema and metadata.

It was necessary to adjust how the PageAnswersPresenter collected the
answers when passing them into the CYA page.

Previously the map function returned a flat array of Presenters.
When there is only one question per page this is fine as the CYA page
uses the label or the legend of the component and does not think about
the page at all.

However for multiple question pages the questions are essentially
grouped by the heading on the page. By the time we get to the CYA page
it makes sense to group the answers by the page heading and present that
back to the user.

In order to achieve this we now collect the page answers grouped by the
page that they are found on, then iterate over those pages and then the
answers checking whether it is necessary to show the heading of the page
or not.