Move _id and _type to page schema
The _id and _type are always required when generating a new page. We may
as well put them as required fields in the page definition as all the
pages eventually inherit from that definition.

Also remove the additionalProperties = false for the checkanswers and
confirmation pages because that flag validates too early. The definition
that relates to properties such as _id and _uuid is in the page
definition and would therefore always fail.

For example, with the confirmation page we define properties such as
heading, body and lede. However the metadata generated will contain
_uuid which is correct, but the additionalProperties flag would fail the
validation before it goes to check the page definition.