Add content page schema and default metadata
This adds the content page schema, default metadata and the template.
There is only meant to be the section heading, heading, lede and body.
So very similar to the start page.

On top of that there will be the ability to add another content
component in a future story.

The continue button remains as the page is meant to be for informative
purposes. Since the public user is not going to be entering any
information this content page will never appear on the check your
answers page nor in the submission itself.