Separate out default text from default metadata
We set the initializer method for the metadata object to now take a
attribute called editor which by default will be false. This is used to
check whether or not to show any default text for any given element.

In the editor we want default text to be shown if the user has not
entered any custom text into a field. However in the runner if the
editor user has not entered any custom text then we do not want the
default text to be shown to the public user.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>