Validate individual pages in before hook
Validation of the whole metadata is not working quite as expected at the
moment. In the short term validate each page against the correct schema.

This is used mainly by the metadata api on every request it receives.

Also add in the page types as references to the base schema instead of
referencing the page definition.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>