Add standalone page schemas, default metadata and templates
This adds the concept of a standalone page. Since they are not part of
the standard flow we did not want to add them to the already existing
pages array. We may want to revisit this decision but for now this is
fine.

By their very definition a standalone page exists independently of form.