Reject non editable attributes from page
There are some additional attributes that we do not want to be editable
when the page is rendered in the editor. add_component and
add_extra_component are fields which the JS writes as hidden fields as
and when a user interacts with the UI. Therefore we don't want them
there on page load.

Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>
Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>