Fix CYA table layout to not include content components
The CYA table was accidentally including content components in the
layout. These should be removed from the components that the answers
presenter gathers.

Also if there was a multiple question we need to close the </dl> tag and
open another <dl> tag when it's the last question for that page.
Otherwise it groups the next question into the same part of the table as
the multiple questions pages questions.