Add upload page to fixture
We also moved the accept and max_size to the validations schema.
We don't need uploadBundle property because the new runner is different
from the classic/legacy form builder.

On the new runner we have errors and validation as separate nodes

   components: [
     {
       errors: {
         "required": "Some custom required message"
       },
       validation: {
         "required": true
       }
     }
   ]

So we are not separating by component type like number_bundle,
   uploadBundle etc

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>