Add upload mechanism to the gem
Cater for upload file hash in user answers

If user is answering a page with upload the gem will call
a method called #upload_file that needs to be implemented by
the mounted apps in order to upload the file. Use the page
answers object to handle the upload in the mounted apps

Do not parse answers as json

Add page upload components method

The idea is to dry up other calls to get all upload components
in the runner

Cater for scenarios on upload

When the user does not upload we should return an empty hash otherwise
an exception would be trow for nil class
When the user is the check your answer the upload file in the user
answer will be a Hash so we should cater for that

Make sure the components are an array when selecting upload

On the start page there are no components so we need to handle
this scenario.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>
Co-authored-by: Tomas Destefi <tomas.destefi@digital.justice.gov.uk>