Move file upload logic to the gem
In order to improve cohesion we have to add the upload behaviour
to the gem and make the mountable apps to define the file upload
adapters they want to have.

By default the presenter offers the OfflineUploadAdapter to use
in case you want to use the simple upload version.

Plus the remove file is a get verb because on the editor,
when we mount the preview feature we don't have the rails ujs
so there is no way to work with `method: :delete` parameter.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>
Co-authored-by: Brendan Butler <brendan.butler@digital.justice.gov.uk>