Add is_not operator example
Add spec for pages inside of the branch

When the user is on the page D it should go to the page F below.
Or if the user is on the C should go to the E page.
This is accomplish by the next default property in the flow json.

A -> B -----> D --> F
   |                ^
   ---> C --> E ----

Evaluate the conditions when the next flow object is a branch

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>