Add base flow object schema
The service flow object uses UUIDs as the string property for either a
page object or a branch object. These UUIDs are obviously dynamic so
thes schema has to allow for that as well.

Adding a base schema which makes use of the `patterProperties` to allow
for dynamic uuid string keys.