Add change your answer implementation
We need to check if the user changed answer.
If user didn't change so we will redirect to CYA page.

If the user changed its answer then we need to see if the page
that the user is answering has any future conditions in the form then
the user will go one page at the time until the end of the form.

Co-authored-by: Natalie Seeto <natalie.seeto@digital.justice.gov.uk>