Add grid and route models
Further details and diagrams can be found in the [PR related to
this](https://github.com/ministryofjustice/fb-metadata-presenter/pull/179).

This adds a couple of models which are used to:

- Traverse all potential routes for a given service
- Build a grid based object which serves the service flow page

The Route model is used to traverse a potential path through a service's
logical flow. It will follow the default next destination for a given
page.

The first Route always traverses from the start page.

When it gets to a branch object it will take the first conditional
destination and follow that through until the end of the form, or until
an exit page is reached along that route.

It will then create additional Route instances which will traverse from
the conditional destination.

Route objects have no knowledge of any other Route objects. They simply
represent a possible path that can be taken through a service.

Each Route object knows the row and column location of where it starts
to traverse from. Since the Route has no knowledge of what is in other
Routes, the Grid object sometimes has to update the starting row for a
given Route.

The Grid model is responsible for taking all the potential Routes
through a service and building out a matrix/grid (array of arrays) that
represents the rows and columns layout of the service flow.

It first builds a matrix based upon the total number of potential routes
by the maximum number of potential columns. The latter is figured out by
taking the Route with the highest starting column number and adding on
the total number of unique flow UUIDs that that Route traversed.

Always starting with the Route that is traversing from the start page,
the Grid first iterates over each UUID and always the highest column
number that that specific object will appear it. By doing this it always
makes sure to put the object in the highest column possible which means
that row Spacers are kept in the right place by default.

Once columns have been added for each flow object, the UUIDs are
iterated over again and if it is the first time that object was seen
then the row is updated. Otherwise it is skipped.

After the above two steps any flow objects in that still have nil value
for their row and column properties are considered detached from the
main flow.

Route rows need to also be udpated when iterating over each branch object.
All the potential destinations that are exiting from a given branch are
looked at and any that already exist in any of the rows above are
ignored. Then the Route objects that have their starting point from any
of the remaining destinations are found and their rows updated in
relation to the current row that the branch object is on. Doing this
prevents any unwanted Spacer objects being left in columns directly
after a branching point.

There is one caveat to the above paragraph and that is when additional
space is required for the conditional questions and answers that are
displayed directly after each branching point.

The UI treats each 'or' expression type for any given conditional as
something that requires it's own line and arrow. Therefore an additional
Spacer object needs to be injected into the corresponding column
directly after a branch if any of it's conditionals is an 'or'. Any
conditionals which are 'and' are still grouped together and therefore do
not need to be separated out in the same way.

Lastly any Spacer objects that are after the last position of a Flow
object for any given column are removed.

The Grid object also will return a flattened ordered flow array
continaing branches and pages and a flattened ordered pages array
containing only pages. These are used for building the Destination lists
in the Change Destination action as well as the Questions list on the
Edit Branch page. Neither of these require Spacer objects, they just
need to be in the correct relative order of the service flow.

We have added a couple of ActiveSupport::Notifications for the Route and
Grid objects. These are here to warn about times when iterating loops
may have continued to be infinite. We calculate the total number of
potential routes as well as the total number of objects that need to be
iterated over and break if those numbers are ever exceeded. We will
monitor this to see if it is appropriate.