Use pointer objects to reference objects in main flow
Should any object in a detached flow already exist in the main service
flow we want to represent that using a Pointer object. This is so the UI
will know to only show the object title in the flow, not the full blown
object.

We also want to stop the visual represenation of the flow at the first
Pointer we come across in any given route.