Add utility methods to the service object
There will be times that we need to check all conditionals in the form
or we need to check all expressions in the form (like when we
delete pages we should check if there is any conditional for that page,
etc).

This commit adds those methods to dry up in the other places where its
used.