Surface start_from in Grid models
There can be n number of detached flows. However we do not want to have
a separate flow for _every_ single UUID that is missing from the main
service flow.

Therefore only create flows starting from UUIDs that don't exist in the
main flow AND that do not exist in any previously created detached
flows.