Allow traversable of objects later in the flow
Previously we were not considering objects that would have appeared
later in the flow if they had been previously traversed. i.e if an objet
is pointed to twice we would only be considering the first time it
appeared in the flow.

This resulted in the accidental detaching of certain objects despite the
fact they were actually linked to.

Update this to only traverse if a route has not been traversed as
opposed to each individual object.