Make space for visually stacked branches
We do not allow branches to be stacked in terms of data integrity but we
do allow them to be visually stacked. In these instances we need to make
space for branch expressions ('and' and 'or' types) so that the frontend
has the ability to display the question and answers in the pages flow.

At the same time this change also takes into account secondary routes
that may or may not link to pages and branches that have been previously
linked to earlier in the flow. In those instances we want to take the
later column, and under some circumstances we want to also update the
row numbers.