Cater for branches being more than one row above objects
In some instances it is possible for a Flow object to be in the same
column as a branch object which is above it. However it might be more
than one row above it.

Make sure there are enough spaces in between for the expressions. At the
same time do not add more spaces than necessary.