Do not allow the overwriting of previously placed Flow objects
Previously the Grid did not check whether an object with a different
uuid existed in the location that is is attempting to place another Flow
object.

Check whether the space is already occupied and if so insert a Spacer
before placing the new object.

This is not currently able to cater for the edge cases which involve the
ordering of destinations exiting from a branch. However it is more
important at this point that _all_ required objects that are linked to
appear on the page instead of becoming detached.