Adjust row number based on branch spacers
When a branch object row is calculated we also calculate the positions
of the required Spacer objects that have to go below it in the grid.

Now we associate those Spacers to a specific object UUID. Doing this
enables us to check whether the objects in the column following a branch
need to be moved to a higher row.

The Spacers for each branch are used to hold the text associated with
any conditional logic question and answer which dictates the visual
flow. If we did not update the row numbers then this can lead to objects
being left on rows further up the page.