Remove excess Pointer objects in detached grids
The rules for building grids which are used in the detached section of
the Editor is a little different from the one used for the main flow.

The main flow usually takes the column number of the object that was
previously positioned, unless it's a checkanswers page or confirmation
page, in which case it will use the latest column number.

For detached grids we want to trim Pointer objects back so each row ends
in the first Pointer it comes across. However in many instances the
object that is being pointed to might be on a completely different row
and either later or earlier in the flow. This is mostly a problem when
it comes to branching point destinations.

In those instances we check whether there should be a Pointer for that
object. If so then we replace the Spacer with a Pointer.

Conversely if there are any rows that only contain Spacers and Pointers,
and none of the Pointers are a branch destination from the column
directly before, then we do not really want that row at all. In these
cases swap any Pointers for Spacers and allow the trim_spacers method to
remove the rows at the end of building the grid.