Set previous flow UUIDs for route objects
This adds previous flow uuids on the route model for any objects that it
traverses. It will also add a conditional uuid should that be
applicable.

For any given route object, any branches that are part of that
particular route, the first conditional will be traversed as normal and
the previous flow uuid and conditional uuid set for that object.

Each additonal branch destination and default next for that branch get
their own route object which will traverse that path and set previous
flow uuids and conditional uuids accordingly.