Set previous uuids for objects in a grid
The route model exposes the previous flow uuid and conditional uuid for
each flow object that is traversed for any given route.

These now need to be set for any given grid, main flow or detached.
Similar rules apply for when flow objects are placed on the grid; the
first flow object takes precedence over any preceeding objects. This is
to say that only the first object that is placed on the grid will have
the ability to be moved as the previous flow uuid and optional
conditional uuid will not be overwritten by any following objects.