Date before and after validations should ignore invalid dates
The date component already has a date validator on it by default which
checks whether the user input is in fact a valid date. Therefore the
date_after and date_before components should not attempt to validate
whether the date is before or after what is configured. Just let the
date validator error surface back to the user.