Format date validation configuration
For the date_after and date_before validations we want to be able to
inject a formatted date into the error message to the user. However we
do not want to do this for the `date` validator itself as the value for
this is a boolean and not a date string.