Add analytics script tags to layout
This adds the script tags for analytics to the head of the application
layout. We initially support Google Analytics at this time. There are
three types:

- UA (Universal Analytics)
- GTM (Google Tag Manager)
- GA4 (Google Analytics 4)

Each one injects a different script into the head. In the case of Google
Tag Manager it also adds a no-script tag to the beginning of the body.